requirements.txt
discord.py==2.2.2
openai==0.27.0
pandas==1.5.3
python-dotenv==1.0.0
tiktoken==0.3.0